18112024_pokemon/requirements.txt
annotated-types==0.7.0
anyio==4.6.2.post1
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
decorator==4.4.2
exceptiongroup==1.2.2
fastapi==0.115.5
h11==0.14.0
httpcore==1.0.7
httpx==0.27.2
idna==3.10
imageio==2.36.0
imageio-ffmpeg==0.5.1
imagekitio==4.1.0
moviepy==1.0.3
numpy==2.0.2
pillow==11.0.0
proglog==0.1.10
pydantic==2.9.2
pydantic_core==2.23.4
python-dotenv==1.0.1
rcs==1.0.16
requests==2.32.3
requests-toolbelt==0.10.1
sniffio==1.3.1
starlette==0.41.2
tqdm==4.67.0
types-Pillow==10.2.0.20240822
typing_extensions==4.12.2
urllib3==1.26.20
uvicorn==0.32.0
